feat: Settings Page aufteilen - Navigation zu Admin-Bereichen
- Add admin navigation section to settings page with links to:
  - Kategorien (/admin/categories)
  - Lagerorte (/admin/locations)
  - Benutzer (/admin/users)
  - Gefrierzeiten (/admin/freeze-times)
- Extract freeze time configuration to new page /admin/freeze-times
- Keep Smart Default settings on settings page
- Update tests for both pages

closes #79
