+++ app/ui/pages/__init__.py
@@ -6,6 +6,7 @@ Importiert alle Pages um sie bei NiceGUI zu registrieren.
 from . import add_item  # noqa: F401
 from . import categories  # noqa: F401
 from . import dashboard  # noqa: F401
+from . import freeze_times  # noqa: F401
 from . import items  # noqa: F401
 from . import locations  # noqa: F401
 from . import login  # noqa: F401
